requirements.txt
anyio==3.6.2
autopep8==2.0.2
click==8.1.3
fastapi==0.92.0
FastAPI-SQLAlchemy==0.2.1
h11==0.14.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
Pygments==2.14.0
rich==13.3.1
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
